classifier/requirements.txt
tiktoken==0.8.0
openai==1.52.2
pandas==2.2.2
scikit-learn==1.5.1
transformers==4.45.2
torch==2.5.0
numpy==1.26.4
accelerate==1.0.1
pyparsing==3.1.2
python-dotenv==1.0.1
pytorch-lightning==2.4.0
tqdm==4.66.5
spacy==3.7.2
en-core-web-trf==3.5.0
spacy[cuda113]